requirements.txt
langchain-huggingface
langchain-groq
langchain-core
python-dotenv
tiktoken
pymongo
numpy
spacy
fastapi
uvicorn
pydantic
beautifulsoup4
python-docx
pandas
PyMuPDF
requests
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]
python-multipart